Isolate ProgressIndexer
This helps keep tested code free of Obsidian deps.
